requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
argilla==1.11.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
commonmark==0.9.1
cryptography==41.0.1
dataclasses-json==0.5.8
decorator==5.1.1
Deprecated==1.2.14
et-xmlfile==1.1.0
faiss-cpu==1.7.4
filetype==1.2.0
flake8==6.0.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.7.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langchain==0.0.210
langchainplus-sdk==0.0.17
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mccabe==0.7.0
mdurl==0.1.2
monotonic==1.6
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
protobuf==4.23.3
pyarrow==12.0.1
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.9
pydeck==0.8.1b0
pyflakes==3.0.1
Pygments==2.15.1
Pympler==1.0.1
pypandoc==1.11
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rich==13.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.17
streamlit==1.23.1
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.7.8
urllib3==2.0.3
validators==0.20.0
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
